requirements.txt
discord.py
spotify
flask
gunicorn